feat(merge-tree): add basic perf tests (#13118)
Adds basic benchmarks for merge-tree operations
