+++ experimental/dds/tree/package.json
@@ -58,7 +58,7 @@
 	},
 	"devDependencies": {
 		"@fluid-internal/stochastic-test-utils": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
-		"@fluid-tools/benchmark": "^0.45.0",
+		"@fluid-tools/benchmark": "^0.46.0",
 		"@fluidframework/build-common": "^1.1.0",
 		"@fluidframework/container-loader": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/container-runtime": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
