+++ packages/dds/matrix/package.json
@@ -81,7 +81,7 @@
 	},
 	"devDependencies": {
 		"@fluid-internal/test-dds-utils": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
-		"@fluid-tools/benchmark": "^0.45.0",
+		"@fluid-tools/benchmark": "^0.46.0",
 		"@fluid-tools/build-cli": "^0.10.0",
 		"@fluidframework/build-common": "^1.1.0",
 		"@fluidframework/build-tools": "^0.10.0",
