+++ packages/test/stochastic-test-utils/package.json
@@ -62,7 +62,7 @@
 		"best-random": "^1.0.0"
 	},
 	"devDependencies": {
-		"@fluid-tools/benchmark": "^0.45.0",
+		"@fluid-tools/benchmark": "^0.46.0",
 		"@fluidframework/build-common": "^1.1.0",
 		"@fluidframework/eslint-config-fluid": "^2.0.0",
 		"@fluidframework/mocha-test-setup": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
