+++ packages/test/test-end-to-end-tests/package.json
@@ -73,7 +73,7 @@
 	},
 	"dependencies": {
 		"@fluid-experimental/sequence-deprecated": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
-		"@fluid-tools/benchmark": "^0.45.0",
+		"@fluid-tools/benchmark": "^0.46.0",
 		"@fluidframework/agent-scheduler": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/aqueduct": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/attributor": ">=2.0.0-internal.3.3.0 <2.0.0-internal.4.0.0",
